Track upstream changes to mockwebserver.
This removes the MockStreamResponse class, replacing it with the
previous MockResponse class. The MockStreamResponse class added
too much to the MockWebServer API for too little benefit: the
BaseResponse class wasn't 'public' but it was required to be public
API for callers to call MockWebServer.enqueue(). This approach is
less object-oriented, but also a lot smaller.

Change-Id: I8496fc3c4dcbfaaf6cd5e97f06d2786b49fa51e1